+++ src/index.html
@@ -67,6 +67,11 @@
           <img class="lightbox__image" />
         </div>
 
+        <button type="button" class="lightbox__btn">
+          <svg class="close-icon">
+            <use href="./img/sprite.svg#close-button"></use>
+          </svg>
+        </button>
       </div>
 
     </section>
@@ -76,13 +81,13 @@
     <div class="pagination container">
 <div class="pagination-style">
       <button class="btn-page previous" data-index="1">&larr;</button>
-      <!-- <span class="btn-dots">...</span> -->
+      <span class="btn-dots js-dots1">...</span>
       <button class="btn-page1 active btn" data-index="1">1</button>
       <button class="btn-page2 btn" data-index="2">2</button>
       <button class="btn-page3 btn" data-index="3">3</button>
       <button class="btn-page4 btn" data-index="4">4</button>
       <button class="btn-page5 btn" data-index="5">5</button>
-      <span class="btn-dots">...</span>
+      <span class="btn-dots js-dots">...</span>
       <button class="btn-page next" data-index="1">&rarr;</button>
     </div>
    </div>
