+++ src/js/fetchMovies.js
@@ -1,89 +1,13 @@
 import refs from './refs';
 import updateMarkup from './updateMarkup';
 import movieCardsTpl from '../template/movieCards.hbs';
+import { currentPage } from './pagination'
 
 import { Spinner } from 'spin.js';
 import opts from './spinner'
 
 
 const apiKey = "api_key=50b81e1c6c3b9e5f74d2015b742ff0b0";
-let currentPage = 1;
-let btns = document.querySelectorAll('.btn');
-
-function setBtnActiveStyle() {
-  let btnEvent = event.target;
-  btns.forEach((el) => el.classList.remove("active"));
-  if (btnEvent.classList.contains("btn")) {
-    btnEvent.classList.add("active")
-  }
-
-}
-
-
-refs.pagination.addEventListener("click", onBtnClick);
-
-function onBtnClick(event) {
-
-  if (event.target.tagName === 'BUTTON') {
-
-    const activeBtn = event.target.dataset.index;
-    console.log('activeBtn', activeBtn);
-    currentPage = Number(activeBtn);
-
-    const btn1 = refs.btnPage1;
-    const btn2 = refs.btnPage2;
-    const btn3 = refs.btnPage3;
-    const btn4 = refs.btnPage4;
-    const btn5 = refs.btnPage5;
-    const previous = refs.previous;
-    const next = refs.next;
-
-    if (event.target.classList.contains("next") && currentPage < 999) {
-      btns.forEach((el) => el.classList.remove("active"))
-      btn1.classList.add("active");
-      next.dataset.index = Number(next.dataset.index) + 5;
-      btn1.textContent = Number(btn1.textContent) + 5;
-      btn2.textContent = Number(btn2.textContent) + 5;
-      btn3.textContent = Number(btn3.textContent) + 5;
-      btn4.textContent = Number(btn4.textContent) + 5;
-      btn5.textContent = Number(btn5.textContent) + 5;
-      btn1.dataset.index = Number(btn1.dataset.index) + 5;
-      btn2.dataset.index = Number(btn2.dataset.index) + 5;
-      btn3.dataset.index = Number(btn3.dataset.index) + 5;
-      btn4.dataset.index = Number(btn4.dataset.index) + 5;
-      btn5.dataset.index = Number(btn5.dataset.index) + 5;
-      currentPage = next.dataset.index;
-    }
-
-    previous.dataset.index = currentPage;
-
-    if (event.target.classList.contains("previous") && currentPage > 5) {
-      next.dataset.index = Number(next.dataset.index) - 5;
-      previous.dataset.index = next.dataset.index
-      btn1.textContent = Number(btn1.textContent) - 5;
-      btn2.textContent = Number(btn2.textContent) - 5;
-      btn3.textContent = Number(btn3.textContent) - 5;
-      btn4.textContent = Number(btn4.textContent) - 5;
-      btn5.textContent = Number(btn5.textContent) - 5;
-      btn1.dataset.index = Number(btn1.dataset.index) - 5;
-      btn2.dataset.index = Number(btn2.dataset.index) - 5;
-      btn3.dataset.index = Number(btn3.dataset.index) - 5;
-      btn4.dataset.index = Number(btn4.dataset.index) - 5;
-      btn5.dataset.index = Number(btn5.dataset.index) - 5;
-      currentPage = previous.dataset.index;
-    }
-    console.log('currentPage>', currentPage);
-    // console.log('previous>', previous.dataset.index);
-    if (event.target.classList.contains("btn")) {
-      setBtnActiveStyle(event)
-    }
-    refs.gallery.innerHTML = '';
-    fetchMovies();
-  }
-}
-
-
-
 
 function fetchMovies() {
 
@@ -92,14 +16,11 @@ function fetchMovies() {
   var target = document.getElementById('gallery');
   var spinner = new Spinner(opts).spin(target);
 
-
-
   return fetch(url)
     .then(response => response.json())
     .then(({ results }) => {
       fetchGenres().then(({ genres }) => {
         updateMovieMarkup(results, genres);
-        // .then(data => updateMarkup(data)
       }).finally(() => {
         spinner.stop();
       });
@@ -111,7 +32,6 @@ function fetchGenres() {
     `https://api.themoviedb.org/3/genre/movie/list?${apiKey}`,
   ).then(response => response.json());
 }
-
 
 
 function updateMovieMarkup(films, genres) {
@@ -130,7 +50,6 @@ function updateMovieMarkup(films, genres) {
     updateMarkup(movie);
   });
 }
-// fetchMovies();
 
 export default fetchMovies;
