+++ src/js/refs.js
@@ -16,6 +16,7 @@ const refs = {
   modalDiv: document.querySelector('.js-lightbox'),
   lightbox: document.querySelector('.lightbox'),
   overlayDiv: document.querySelector('.lightbox__overlay'),
+  btnClose: document.querySelector('.lightbox__btn'),
   modalDivContent: document.querySelector('.lightbox__content'),
   libraryButton: document.querySelector('.library-ref'),
   container: document.getElementById('.container'),
@@ -32,6 +33,8 @@ const refs = {
   btnPage5: document.querySelector('.btn-page5'),
   previous: document.querySelector('.previous'),
   next: document.querySelector('.next'),
+  dots1: document.querySelector('.js-dots1'),
+  dots2: document.querySelector('.js-dots2'),
 
   queueBtn: searchBySelector('#watched-queue-btn'),
   watchedBtn: searchBySelector('#watched-film-btn'),
