+++ src/scss/layout/_modal_cards.scss
@@ -207,3 +207,30 @@
 .lightbox__button:focus {
   background-color: rgba(0, 0, 0, 0.5);
 }
+
+.lightbox__btn {
+  color: #f7f7f7;
+}
+
+.lightbox__btn {
+  position: absolute;
+  top: 8px;
+  right: 8px;
+  width: 30px;
+  height: 30px;
+  border: none;
+  transition: border-color 250ms cubic-bezier(0.4, 0, 0.2, 1);
+  outline: 2px solid #ff6b08;
+}
+
+.close-icon {
+  width: 100%;
+  height: 100%;
+  background-color: #f7f7f7;
+  fill: #ff6b08;
+  transition: fill 250ms cubic-bezier(0.4, 0, 0.2, 1);
+}
+.lightbox__btn:hover .close-icon {
+  fill: #f7f7f7;
+  background-color: #ff6b08;
+}
