+++ src/scss/layout/_pagination.scss
@@ -28,10 +28,8 @@
   @media screen and (max-width: 640px) {
     width: 30px;
     height: 30px;
+  }
 }
-}
-
-
 
 .btn-page1:hover,
 .btn-page2:hover,
@@ -45,18 +43,16 @@
 }
 
 .btn-dots {
-  display: inline-block;
   width: 30px;
   height: 27px;
   vertical-align: middle;
   margin-right: 15px;
   @media screen and (max-width: 640px) {
     display: none;
-}
+  }
 }
 
 .active {
   background: #ff6b08;
   color: #ffff;
 }
-
