+++ src/scss/layout/pagination.scss
@@ -44,7 +44,6 @@
 }
 
 .btn-dots {
-  display: inline-block;
   width: 40px;
   height: 33px;
   vertical-align: middle;
